rename fourd to bottom up, add metric and alignment, change name manner
