+++ configs/4dag/shelf_config/eval_keypoints3d.py
@@ -1,4 +1,4 @@
-type = 'FourDAssociationEvaluation'
+type = 'BottomUpAssociationEvaluation'
 
 __data_root__ = '../data/Shelf'
 __meta_path__ = __data_root__ + '/xrmocap_meta_testset'
@@ -14,48 +14,48 @@ selected_limbs_name = [
 additional_limbs_names = [['jaw', 'headtop']]
 
 associator = dict(
-    type='FourdAssociator',
-    m_minAsgnCnt=5,
+    type='BottomUpAssociator',
+    min_asgn_cnt=5,
     m_filter = False,
     triangulator=dict(
         type='FourDAGTriangulator',
-        m_activeRate=0.1,
-        m_minTrackJCnt=5,
-        m_boneCapacity=100,
-        m_wBone3d=1.0,
-        m_wSquareShape=1e-2,
-        m_shapeMaxIter=5,
-        m_wJ3d=1.0,
-        m_wRegularPose=1e-3,
-        m_poseMaxIter=20,
-        m_wJ2d=1e-5,
-        m_wTemporalTrans=1e-1,
-        m_wTemporalPose=1e-2,
-        m_minTriangulateJCnt=15,
-        m_initActive=0.9,
-        m_triangulateThresh=0.05,
+        active_rate=0.1,
+        min_track_cnt=5,
+        bone_capacity=100,
+        w_bone3d=1.0,
+        w_square_shape=1e-2,
+        shape_max_iter=5,
+        w_joint3d=1.0,
+        w_regular_pose=1e-3,
+        pose_max_iter=20,
+        w_joint2d=1e-5,
+        w_temporal_trans=1e-1,
+        w_temporal_pose=1e-2,
+        min_triangulate_cnt=15,
+        init_active=0.9,
+        triangulate_thresh=0.05,
         logger=logger,
     ),
     fourd_matching=dict(
         type='FourdMatching',
-        m_maxEpiDist=0.15,
-        m_maxTempDist=0.2,
-        m_wEpi=2,
-        m_wTemp=2,
-        m_wView= 2,
-        m_wPaf= 1 ,
-        m_wHier= 0.5,
-        m_cViewCnt= 1.5,
-        m_minCheckCnt= 1,
-        m_minAsgnCnt= 5 ,
-        m_normalizeEdges = True,
+        max_epi_dist=0.15,
+        max_temp_dist=0.2,
+        w_epi=2,
+        w_temp=2,
+        w_view= 2,
+        w_paf= 1 ,
+        w_hier= 0.5,
+        c_view_cnt= 1.5,
+        min_check_cnt= 1,
+        min_asgn_cnt= 5 ,
+        normalize_edges = True,
         logger=logger,
     ),
     logger=logger,
 )
 
 dataset = dict(
-    type='FourDDataset',
+    type='BottomUpMviewMpersonDataset',
     data_root=__data_root__,
     img_pipeline=[
         dict(type='LoadImagePIL'),
@@ -65,7 +65,6 @@ dataset = dict(
     meta_path=__meta_path__,
     test_mode=True,
     shuffled=False,
-    kps2d_convention=pred_kps3d_convention,
     gt_kps3d_convention='campus',
     cam_world2cam=True,
 )
