+++ configs/modules/ops/bottom_up_association/bottom_up_association.py
@@ -1,0 +1,35 @@
+type='BottomUpAssociator'
+min_asgn_cnt=5
+m_filter = False
+triangulator=dict(
+    type='FourDAGTriangulator',
+    active_rate=0.1,
+    min_track_cnt=5,
+    bone_capacity=100,
+    w_bone3d=1.0,
+    w_square_shape=1e-2,
+    shape_max_iter=5,
+    w_joint3d=1.0,
+    w_regular_pose=1e-3,
+    pose_max_iter=20,
+    w_joint2d=1e-5,
+    w_temporal_trans=1e-1,
+    w_temporal_pose=1e-2,
+    min_triangulate_cnt=15,
+    init_active=0.9,
+    triangulate_thresh=0.05,
+)
+fourd_matching=dict(
+    type='FourdMatching',
+    max_epi_dist=0.15,
+    max_temp_dist=0.2,
+    w_epi=2,
+    w_temp=2,
+    w_view= 2,
+    w_paf= 1 ,
+    w_hier= 0.5,
+    c_view_cnt= 1.5,
+    min_check_cnt= 1,
+    min_asgn_cnt= 5 ,
+    normalize_edges = True,
+)
