+++ xrmocap/core/evaluation/builder.py
@@ -2,14 +2,14 @@ from mmcv.utils import Registry
 
 from .mvp_evaluation import MVPEvaluation
 from .top_down_association_evaluation import TopDownAssociationEvaluation
-from .fourd_association_evaluation import FourDAssociationEvaluation
+from .bottom_up_association_evaluation import BottomUpAssociationEvaluation
 EVALUATION = Registry('evaluation')
 
 EVALUATION.register_module(
     name='TopDownAssociationEvaluation', module=TopDownAssociationEvaluation)
 EVALUATION.register_module(name='MVPEvaluation', module=MVPEvaluation)
 EVALUATION.register_module(
-    name='FourDAssociationEvaluation', module=FourDAssociationEvaluation)
+    name='BottomUpAssociationEvaluation', module=BottomUpAssociationEvaluation)
 
 
 def build_evaluation(cfg):
