+++ xrmocap/data/dataset/builder.py
@@ -4,7 +4,7 @@ from mmcv.utils import Registry
 from .base_dataset import BaseDataset
 from .mview_mperson_dataset import MviewMpersonDataset
 from .mvp_dataset import MVPDataset
-from .fourd_dataset import FourDDataset
+from .bottom_up_mview_mperson_dataset import BottomUpMviewMpersonDataset
 
 # yapf: enable
 
@@ -13,7 +13,7 @@ DATASETS.register_module(
     name='MviewMpersonDataset', module=MviewMpersonDataset)
 DATASETS.register_module(name='MVPDataset', module=MVPDataset)
 DATASETS.register_module(
-    name='FourDDataset', module=FourDDataset)
+    name='BottomUpMviewMpersonDataset', module=BottomUpMviewMpersonDataset)
 
 
 def build_dataset(cfg) -> BaseDataset:
